Move to openshift-gce-devel account; make gcp specific
